16/10 add new read me

+++ src/utils/file_remove.py
@@ -1,5 +1,7 @@
 import os
+
 from typing import List
+
 from src.logs import Log
 
 
@@ -17,8 +19,10 @@ def delete_local_files(files_path: List) -> None:
     Log.info(f'validation delete_local_files started ')
     for file_path in files_path:
         try:
-            os.remove(file_path)
+            if os.path.exists(file_path):
+                os.remove(file_path)
         except OSError as e:
             error_message: str = "An error occurred while deleting local file : " + str(e)
             Log.error(f'validation delete_local_files failed : {error_message}')
             raise ValueError(error_message)
+    Log.info(f'done activate - validation delete_local_files  ')
